requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
openai
SpeechRecognition
pyaudio
